fix:chromadb delete_by_ids error Close #1021

+++ dbgpt/storage/vector_store/chroma_store.py
@@ -113,8 +113,11 @@ class ChromaStore(VectorStoreBase):
 
     def delete_by_ids(self, ids):
         logger.info(f"begin delete chroma ids...")
-        collection = self.vector_store_client._collection
-        collection.delete(ids=ids)
+        ids = ids.split(",")
+        if len(ids) > 0:
+            collection = self.vector_store_client._collection
+            collection.delete(ids=ids)
+        logger.info(f"delete chroma by ids success...")
 
     def _clean_persist_folder(self):
         for root, dirs, files in os.walk(self.persist_dir, topdown=False):
